+++ nemo/collections/tts/modules/__init__.py
@@ -16,3 +16,4 @@ from nemo.collections.tts.modules.tacotron2 import Decoder as Taco2Decoder
 from nemo.collections.tts.modules.tacotron2 import Encoder as Taco2Encoder
 from nemo.collections.tts.modules.tacotron2 import Postnet as Taco2Postnet
 from nemo.collections.tts.modules.waveglow import WaveGlowModule
+import nemo.collections.tts.modules.adapters
