+++ nemo/collections/tts/modules/fastpitch.py
@@ -42,12 +42,8 @@
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-from typing import List, Optional
-
 import torch
-from omegaconf import DictConfig
 
-from nemo.collections.asr.parts.utils import adapter_utils
 from nemo.collections.tts.modules.submodules import ConditionalInput, ConditionalLayerNorm
 from nemo.collections.tts.parts.utils.helpers import binarize_attention_parallel, regulate_len
 from nemo.core.classes import NeuralModule, adapter_mixins, typecheck
@@ -151,39 +147,6 @@ class TemporalPredictor(NeuralModule):
         out = out.transpose(1, 2)
         out = self.fc(out) * enc_mask
         return out.squeeze(-1)
-
-
-class TemporalPredictorAdapter(TemporalPredictor, adapter_mixins.AdapterModuleMixin):
-    """ Inherit from TemporalPredictor and add support for adapter"""
-
-    def add_adapter(self, name: str, cfg: dict):
-        cfg = self._update_adapter_cfg_input_dim(cfg)
-        for conv_layer in self.layers:  # type: adapter_mixins.AdapterModuleMixin
-            conv_layer.add_adapter(name, cfg)
-
-    def is_adapter_available(self) -> bool:
-        return any([conv_layer.is_adapter_available() for conv_layer in self.layers])
-
-    def set_enabled_adapters(self, name: Optional[str] = None, enabled: bool = True):
-        for conv_layer in self.layers:  # type: adapter_mixins.AdapterModuleMixin
-            conv_layer.set_enabled_adapters(name=name, enabled=enabled)
-
-    def get_enabled_adapters(self) -> List[str]:
-        names = set([])
-        for conv_layer in self.layers:  # type: adapter_mixins.AdapterModuleMixin
-            names.update(conv_layer.get_enabled_adapters())
-
-        names = sorted(list(names))
-        return names
-
-    def _update_adapter_cfg_input_dim(self, cfg: DictConfig):
-        cfg = adapter_utils.update_adapter_cfg_input_dim(self, cfg, module_dim=self.filter_size)
-        return cfg
-
-
-"""Register any additional information"""
-if adapter_mixins.get_registered_adapter(TemporalPredictor) is None:
-    adapter_mixins.register_adapter(base_class=TemporalPredictor, adapter_class=TemporalPredictorAdapter)
 
 
 class FastPitchModule(NeuralModule, adapter_mixins.AdapterModuleMixin):
