+++ nemo/collections/tts/modules/transformer.py
@@ -11,14 +11,12 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import List, Optional
+from typing import Optional
 
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
-from omegaconf import DictConfig
 
-from nemo.collections.asr.parts.utils import adapter_utils
 from nemo.collections.tts.modules.submodules import ConditionalInput, ConditionalLayerNorm, LinearNorm
 from nemo.collections.tts.parts.utils.helpers import get_mask_from_lengths
 from nemo.core.classes import NeuralModule, adapter_mixins, typecheck
@@ -165,7 +163,7 @@ class MultiHeadAttn(nn.Module):
 
 
 class TransformerLayer(nn.Module, adapter_mixins.AdapterModuleMixin):
-    def __init__(self, n_head, d_model, d_head, d_inner, kernel_size, dropout, condition_types, **kwargs):
+    def __init__(self, n_head, d_model, d_head, d_inner, kernel_size, dropout, condition_types=[], **kwargs):
         super(TransformerLayer, self).__init__()
 
         self.dec_attn = MultiHeadAttn(n_head, d_model, d_head, dropout, condition_types=condition_types, **kwargs)
@@ -181,7 +179,8 @@ class TransformerLayer(nn.Module, adapter_mixins.AdapterModuleMixin):
 
         if self.is_adapter_available():
             output = self.forward_enabled_adapters(output)
-
+            output *= mask
+            
         return output
 
 
@@ -351,47 +350,3 @@ class FFTransformer(nn.Module):
 
         out = self.dense(out).transpose(1, 2)
         return out
-
-
-class FFTransformerDecoderAdapter(FFTransformerDecoder, adapter_mixins.AdapterModuleMixin):
-    """ Inherit from FFTransformerDecoder and add support for adapter"""
-
-    def add_adapter(self, name: str, cfg: dict):
-        cfg = self._update_adapter_cfg_input_dim(cfg)
-        for FFT_layer in self.layers:  # type: adapter_mixins.AdapterModuleMixin
-            FFT_layer.add_adapter(name, cfg)
-
-    def is_adapter_available(self) -> bool:
-        return any([FFT_layer.is_adapter_available() for FFT_layer in self.layers])
-
-    def set_enabled_adapters(self, name: Optional[str] = None, enabled: bool = True):
-        for FFT_layer in self.layers:  # type: adapter_mixins.AdapterModuleMixin
-            FFT_layer.set_enabled_adapters(name=name, enabled=enabled)
-
-    def get_enabled_adapters(self) -> List[str]:
-        names = set([])
-        for FFT_layer in self.layers:  # type: adapter_mixins.AdapterModuleMixin
-            names.update(FFT_layer.get_enabled_adapters())
-
-        names = sorted(list(names))
-        return names
-
-    def _update_adapter_cfg_input_dim(self, cfg: DictConfig):
-        cfg = adapter_utils.update_adapter_cfg_input_dim(self, cfg, module_dim=self.d_model)
-        return cfg
-
-
-class FFTransformerEncoderAdapter(
-    FFTransformerDecoderAdapter, FFTransformerEncoder, adapter_mixins.AdapterModuleMixin
-):
-    """ Inherit from FFTransformerEncoder and add support for adapter"""
-
-    pass
-
-
-"""Register any additional information"""
-if adapter_mixins.get_registered_adapter(FFTransformerEncoder) is None:
-    adapter_mixins.register_adapter(base_class=FFTransformerEncoder, adapter_class=FFTransformerEncoderAdapter)
-
-if adapter_mixins.get_registered_adapter(FFTransformerDecoder) is None:
-    adapter_mixins.register_adapter(base_class=FFTransformerDecoder, adapter_class=FFTransformerDecoderAdapter)
